+++ ejercicioEvaluacion/ejercicioExamen/clases_y_atributos/ICalculable.cs
@@ -9,7 +9,7 @@ namespace ejercicioExamen.clases_y_atributos
     public interface ICalculable
     {
         //Definimos un método para implementarlo en Empleado
-        decimal CalcularSalario();
+        double CalcularSalario();
 
         
     }
